refactor(apple): simplify launch/login managers
